requirements.txt.txt
python-telegram-bot
requests
python-nmap
Flask
gunicorn